moved pagination TOs to basic module as they are not JPA specific
