+++ samples/core/src/main/java/io/oasp/gastronomy/restaurant/general/logic/base/AbstractUc.java
@@ -1,7 +1,7 @@
 package io.oasp.gastronomy.restaurant.general.logic.base;
 
 import io.oasp.gastronomy.restaurant.general.common.base.AbstractBeanMapperSupport;
-import io.oasp.module.jpa.common.api.to.PaginatedListTo;
+import io.oasp.module.basic.common.api.to.PaginatedListTo;
 
 import java.util.ArrayList;
 import java.util.Collection;
