+++ samples/core/src/main/java/io/oasp/gastronomy/restaurant/salesmanagement/logic/api/usecase/UcFindOrder.java
@@ -3,7 +3,7 @@ package io.oasp.gastronomy.restaurant.salesmanagement.logic.api.usecase;
 import io.oasp.gastronomy.restaurant.salesmanagement.logic.api.to.OrderCto;
 import io.oasp.gastronomy.restaurant.salesmanagement.logic.api.to.OrderEto;
 import io.oasp.gastronomy.restaurant.salesmanagement.logic.api.to.OrderSearchCriteriaTo;
-import io.oasp.module.jpa.common.api.to.PaginatedListTo;
+import io.oasp.module.basic.common.api.to.PaginatedListTo;
 
 import java.util.List;
