+++ samples/core/src/main/java/io/oasp/gastronomy/restaurant/staffmanagement/logic/api/Staffmanagement.java
@@ -2,7 +2,7 @@ package io.oasp.gastronomy.restaurant.staffmanagement.logic.api;
 
 import io.oasp.gastronomy.restaurant.staffmanagement.logic.api.to.StaffMemberEto;
 import io.oasp.gastronomy.restaurant.staffmanagement.logic.api.to.StaffMemberSearchCriteriaTo;
-import io.oasp.module.jpa.common.api.to.PaginatedListTo;
+import io.oasp.module.basic.common.api.to.PaginatedListTo;
 
 import java.util.List;
