+++ samples/core/src/main/java/io/oasp/gastronomy/restaurant/staffmanagement/logic/api/to/StaffMemberSearchCriteriaTo.java
@@ -1,7 +1,7 @@
 package io.oasp.gastronomy.restaurant.staffmanagement.logic.api.to;
 
 import io.oasp.gastronomy.restaurant.general.common.api.datatype.Role;
-import io.oasp.module.jpa.common.api.to.SearchCriteriaTo;
+import io.oasp.module.basic.common.api.to.SearchCriteriaTo;
 
 /**
  * This is the {@link SearchCriteriaTo search criteria} {@link net.sf.mmm.util.transferobject.api.TransferObject TO}
