+++ samples/core/src/main/java/io/oasp/gastronomy/restaurant/tablemanagement/dataaccess/impl/dao/TableDaoImpl.java
@@ -7,9 +7,9 @@ import io.oasp.gastronomy.restaurant.tablemanagement.common.api.datatype.TableSt
 import io.oasp.gastronomy.restaurant.tablemanagement.dataaccess.api.TableEntity;
 import io.oasp.gastronomy.restaurant.tablemanagement.dataaccess.api.dao.TableDao;
 import io.oasp.gastronomy.restaurant.tablemanagement.logic.api.to.TableSearchCriteriaTo;
-import io.oasp.module.jpa.common.api.to.OrderByTo;
-import io.oasp.module.jpa.common.api.to.OrderDirection;
-import io.oasp.module.jpa.common.api.to.PaginatedListTo;
+import io.oasp.module.basic.common.api.to.OrderByTo;
+import io.oasp.module.basic.common.api.to.OrderDirection;
+import io.oasp.module.basic.common.api.to.PaginatedListTo;
 
 import java.util.List;
